feat: Validate ConfigValue before setting it
This changes all ConfigValue stuff to work with pointers since we need
to mutate the object when doing things like setting values.

It adds a Valid method to ConfigValue that will validate the various
fields based on its ValueType.

Finally update the configvalues.Service to call this new method when
setting a configuration value. Also use the Service for creation of
ConfigValues for this same reason.