feat: Refocus UI on it's primary function, add ability to set new config values
Still need to add editing / updating of existing config values but the
primary goal of this change was to redesign the UI such that it is
focussed more around the task of managing environment configuration.

As such it's now much easier to focus on that task. The environment map
/ tree feature has been removed but will be added separately at a later
date.