feat: Add the ability to edit existing configuration on an environment
As part of this I've also removed the confusing "add inherited values"
behaviour and instead just let them be edited as if they exist on the
environment. This is the far less confusing behaviour for a user and
reduces code complexity as well so win-win.

This obviously begets a change to the canAdd behaviour where we now need
to fetch the ConfigKeys from the API to determine if there are any keys
which aren't configured or inherited. So that functionality can be
tested I've added an unconfigured config key to the seed.go script.