feat: Implement an authn/z system
This has been built in a way that allows for adding additional
authentication / authorization systems in the future such as LDAP or
what have you but to begin with it's basically going to just support
in-built postgres.

There is definitely some room for refactoring and exposing some new
options via the UI / disabling options which won't work based on your
permissions. But it's functional and in the interest of getting
something merged it's definitely "MVP" ready.