feat: Bulk edit and update config values
This required a new API for setting config values as a list. But it
simplifies the frontend code a lot to do it this way.

The API as implemented is quite inefficient and we should add a new
configvalues.Repository code path for doing this bulk update instead of
how it's currently working. Because for large configurations (100+ keys)
it would probably be quite slow.

Also this adds the ability to filter the configuration screen both while
editing and not.

This bulk edit functionality is also important for future updates when
we add webhooks and such since we wouldn't want to fire an event for
every configuration value but instead all of them at once (think when
updating config maps and restarting k8s deployments) or else we'd thrash
deployments in some scenarios.